Moving resources to standard folders.
